requirements.txt
streamlit
PyMuPDF
python-docx
pdfplumber
sentence-transformers
chromadb
requests
